+++ src/XRP/xrp-client.ts
@@ -63,6 +63,8 @@ export default class XrpClient implements XrpClientInterface {
   }
 
   /**
+   * @deprecated Use method `sendXrp` instead.
+   *
    * Send the given amount of XRP from the source wallet to the destination address.
    *
    * @param amount A `BigInteger`, number or numeric string representing the number of drops to send.
@@ -75,7 +77,47 @@ export default class XrpClient implements XrpClientInterface {
     destination: string,
     sender: Wallet,
   ): Promise<string> {
-    return this.sendWithDetails({
+    const transactionResult = await this.sendXrpWithDetails({
+      amount,
+      destination,
+      sender,
+    })
+    return transactionResult.hash
+  }
+
+  /**
+   * @deprecated Use method `sendXrpWithDetails` instead.
+   *
+   * Send the given amount of XRP from the source wallet to the destination PayID, allowing
+   * for additional details to be specified for use with supplementary features of the XRP
+   * ledger.
+   *
+   * @param sendXrpDetails - a wrapper object containing details for constructing a transaction.
+   * @returns A promise which resolves to a string representing the hash of the submitted transaction.
+   */
+  public async sendWithDetails(
+    sendXrpDetails: SendXrpDetails,
+  ): Promise<string> {
+    const transactionResult = await this.decoratedClient.sendXrpWithDetails(
+      sendXrpDetails,
+    )
+    return transactionResult.hash
+  }
+
+  /**
+   * Send the given amount of XRP from the source wallet to the destination address.
+   *
+   * @param amount A `BigInteger`, number or numeric string representing the number of drops to send.
+   * @param destination A destination address to send the drops to.
+   * @param sender The wallet that XRP will be sent from and which will sign the request.
+   * @returns A promise which resolves to a TransactionResult representing the final outcome of the submitted transaction.
+   */
+  public async sendXrp(
+    amount: BigInteger | number | string,
+    destination: string,
+    sender: Wallet,
+  ): Promise<TransactionResult> {
+    return this.sendXrpWithDetails({
       amount,
       destination,
       sender,
@@ -88,12 +130,12 @@ export default class XrpClient implements XrpClientInterface {
    * ledger.
    *
    * @param sendXrpDetails - a wrapper object containing details for constructing a transaction.
-   * @returns A promise which resolves to a string representing the hash of the submitted transaction.
+   * @returns A promise which resolves to a TransactionResult representing the final outcome of the submitted transaction.
    */
-  public async sendWithDetails(
+  public async sendXrpWithDetails(
     sendXrpDetails: SendXrpDetails,
-  ): Promise<string> {
-    return this.decoratedClient.sendWithDetails(sendXrpDetails)
+  ): Promise<TransactionResult> {
+    return this.decoratedClient.sendXrpWithDetails(sendXrpDetails)
   }
 
   /**