+++ CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Method `sendXrp` of class `XrpClient` replaces now deprecated `send`. `sendXrp` returns a `TransactionResult` that
+  contains more granular information about the final outcome of a submitted transaction.
+
+- Method `sendXrpWithDetails` of class `XrpClient` replaces now deprecated `sendWithDetails`.  `sendXrpWithDetails` returns
+  a `TransactionResult` that contains more granular infomration about the final outcome of a submitted transaction.
+
+### Deprecated
+
+- Method `send` of class `XrpClient` is deprecated and will be removed in two release cycles.
+  Use `sendXrp` instead.
+
+- Method `sendWithDetails` of class `XrpClient` is deprecated and will be removed in two release cycles.
+  Use `sendXrpWithDetails` instead.
+
 - A new method `authorizeSendingAccount` is added to `XrpClient` which authorizes
   senders to send XRP to the specified XRPL account.
   (See https://xrpl.org/depositpreauth.html)